+++ packages/tools/src/utilities/math/polyline/index.ts
@@ -9,6 +9,7 @@ import getNormal3 from './getNormal3';
 import getNormal2 from './getNormal2';
 import { mergePolylines, subtractPolylines } from './combinePolyline';
 import intersectPolyline from './intersectPolyline';
+import decimate from './decimate';
 import getFirstLineSegmentIntersectionIndexes from './getFirstLineSegmentIntersectionIndexes';
 import getLineSegmentIntersectionsIndexes from './getLineSegmentIntersectionsIndexes';
 import getLineSegmentIntersectionsCoordinates from './getLineSegmentIntersectionsCoordinates';
@@ -30,6 +31,7 @@ export {
   getNormal3,
   getNormal2,
   intersectPolyline,
+  decimate,
   getFirstLineSegmentIntersectionIndexes,
   getLineSegmentIntersectionsIndexes,
   getLineSegmentIntersectionsCoordinates,